Split AppleReposVIewController into view and viewcontroller
