+++ Challange-1-iOS/Challange-1-iOS/Presentation/AvatarsListScreen/AvatarsListView.swift
@@ -52,6 +52,7 @@ class AvatarsListView: BaseGenericView {
         ])
     }
 }
+
 /*
 extension AvatarsListView: UICollectionViewDelegateFlowLayout {
     func collectionView(_ collectionView: UICollectionView,
@@ -67,18 +68,6 @@ extension AvatarsListView: UICollectionViewDelegateFlowLayout {
         guard let layout = collectionViewLayout as? UICollectionViewFlowLayout else { return .zero }
         let widthPerItem = collectionView.frame.width / 3 - layout.minimumInteritemSpacing
         return CGSize(width: widthPerItem - 8, height: widthPerItem)
-    }
-
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let avatar = self.avatars[indexPath.row]
-        let message: String = "Are you sure that you really want to delete \(avatar.login)?"
-        let alert = UIAlertController(title: "Deleting \(avatar.login)...", message: message, preferredStyle: .alert)
-
-        alert.addAction(UIAlertAction(title: "Cancel", style: .default))
-        alert.addAction(UIAlertAction(title: "Delete", style: .destructive, handler: { (_: UIAlertAction) in
-            self.viewModel?.deleteAvatar(avatar: avatar, at: indexPath.row)
-        }))
-        self.present(alert, animated: true, completion: nil)
     }
 }
 */
