+++ Challange-1-iOS/Challange-1-iOS/Presentation/EmojisListScreen/EmojisListViewController.swift
@@ -12,15 +12,6 @@ class EmojisListViewController: BaseGenericViewController<EmojisListView>, Coord
     var emojisList: [Emoji] = []
     var viewModel: EmojiViewModel?
 
-    init() {
-
-        super.init(nibName: nil, bundle: nil)
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
     override func viewDidLoad() {
         super.viewDidLoad()
         title = "Emojis List"
